Add bin/tec_util for use with modulefile
This update was motivated by the fact that the set-alias command
we were using in the modulefile is not exported to subprocesses. As
a workaround, an actual script was created that invokes tec_util via
"python -m"; this script is then added to the envrionments by adding
the bin directory to the PATH variable, which *is* exported to
subprocesses.